+++ blueprint/implementations/billing/base/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @forklaunch/implementation-billing-base
 
+## 0.8.6
+
+### Patch Changes
+
+- update framework package versions
+- Updated dependencies
+  - @forklaunch/interfaces-billing@0.8.6
+
 ## 0.8.5
 
 ### Patch Changes